@@ -24,77 +24,74 @@ public class BitChoiceGroupTests : BunitTestContext
        DataRow(Visual.Material, false, true),
        DataRow(Visual.Material, false, false)
    ]
-    public void BitChoiceGroupShouldRespectIsEnabled(Visual visual, bool groupIsEnabled, bool optionIsEnabled)
+    public void BitChoiceGroupShouldRespectIsEnabled(Visual visual, bool isEnabled, bool optionIsEnabled)
     {
-        var component = RenderComponent<BitChoiceGroup>(
-            parameters =>
+        var component = RenderComponent<BitChoiceGroup>(parameters =>
+        {
+            parameters.AddCascadingValue(visual);
+            parameters.Add(p => p.IsEnabled, isEnabled);
+            parameters.Add(p => p.Options, new List<BitChoiceGroupOption>
             {
-                parameters.AddCascadingValue(visual);
-                parameters.Add(p => p.IsEnabled, groupIsEnabled);
-                parameters.Add(p => p.Options, new List<BitChoiceGroupOption>
+                new()
                 {
-                    new()
-                    {
-                        Value = "key1",
-                        IsEnabled = optionIsEnabled
-                    }
-                });
+                    Value = "key1",
+                    IsEnabled = optionIsEnabled
+                }
             });
-
-        var bitChoiceGroupOption = component.Find(".bit-chgo");
+        });
 
         var visualClass = visual == Visual.Cupertino ? "cupertino" : visual == Visual.Material ? "material" : "fluent";
+        var isEnabledClass = isEnabled ? "enabled" : "disabled";
 
-        if (groupIsEnabled is false || optionIsEnabled is false)
+        var bitChoiceGroup = component.Find(".bit-chg");
+        var bitChoiceGroupOption = component.Find(".bit-chgo");
+
+        Assert.IsTrue(bitChoiceGroup.ClassList.Contains($"bit-chg-{isEnabledClass}-{visualClass}"));
+
+        if (isEnabled is false)
         {
-            Assert.IsTrue(bitChoiceGroupOption.ClassList.Contains($"bit-chgo-disabled-{visualClass}"));
-        }
-        else
-        {
-            Assert.IsFalse(bitChoiceGroupOption.ClassList.Contains($"bit-chgo-disabled-{visualClass}"));
+            Assert.IsTrue(bitChoiceGroupOption.ClassList.Contains($"bit-chgo-disabled"));
         }
     }
 
     [DataTestMethod, DataRow("key1")]
     public void BitChoiceGroupRespectDafaultValue(string defaultValue)
     {
-        var component = RenderComponent<BitChoiceGroup>(
-           parameters =>
-           {
-               parameters.Add(p => p.DefaultValue, defaultValue);
-               parameters.Add(p => p.Options, new List<BitChoiceGroupOption>
-               {
-                    new()
-                    {
-                        Value = "key1"
-                    },
-                    new()
-                    {
-                        Value = "key2"
-                    }
-               });
-           });
+        var component = RenderComponent<BitChoiceGroup>(parameters =>
+        {
+            parameters.Add(p => p.DefaultValue, defaultValue);
+            parameters.Add(p => p.Options, new List<BitChoiceGroupOption>
+            {
+                new()
+                {
+                    Value = "key1"
+                },
+                new()
+                {
+                    Value = "key2"
+                }
+            });
+        });
 
         var bitChoiceGroupOption = component.Find(".bit-chgo");
 
-        Assert.IsTrue(bitChoiceGroupOption.ClassList.Contains($"bit-chgo-checked-fluent"));
+        Assert.IsTrue(bitChoiceGroupOption.ClassList.Contains($"bit-chgo-checked"));
     }
 
     [DataTestMethod, DataRow("ChoiceGroupName")]
     public void BitChoiceGroupShouldGiveNameBitChoiceGroupOptions(string choiceGroupName)
     {
-        var component = RenderComponent<BitChoiceGroup>(
-             parameters =>
-             {
-                 parameters.Add(p => p.Name, choiceGroupName);
-                 parameters.Add(p => p.Options, new List<BitChoiceGroupOption>
-                 {
-                    new()
-                    {
-                        Value = "key1"
-                    }
-                 });
-             });
+        var component = RenderComponent<BitChoiceGroup>(parameters =>
+        {
+            parameters.Add(p => p.Name, choiceGroupName);
+            parameters.Add(p => p.Options, new List<BitChoiceGroupOption>
+            {
+                new()
+                {
+                    Value = "key1"
+                }
+            });
+        });
 
         var bitChoiceGroupOption = component.Find(".bit-chgo-input");
 
@@ -105,13 +102,12 @@ public class BitChoiceGroupTests : BunitTestContext
     [DataTestMethod, DataRow("this is label")]
     public void BitChoiceGroupShouldTakeLabel(string label)
     {
-        var component = RenderComponent<BitChoiceGroup>(
-              parameters =>
-              {
-                  parameters.Add(p => p.Label, label);
-              });
+        var component = RenderComponent<BitChoiceGroup>(parameters =>
+        {
+            parameters.Add(p => p.Label, label);
+        });
 
-        var bitChoiceGroupLabel = component.Find(".bit-chg-label-fluent");
+        var bitChoiceGroupLabel = component.Find(".bit-chg-label");
 
         Assert.AreEqual(label, bitChoiceGroupLabel.TextContent);
     }
@@ -119,23 +115,22 @@ public class BitChoiceGroupTests : BunitTestContext
     [DataTestMethod, DataRow("https://picsum.photos/100", "this is alt", 50, 50)]
     public void BitChoiceGroupShouldRespectImage(string imageSrc, string imageAlt, int width, int height)
     {
-        var component = RenderComponent<BitChoiceGroup>(
-             parameters =>
-             {
-                 parameters.Add(p => p.Options, new List<BitChoiceGroupOption>
-                 {
-                    new()
-                    {
-                        Value = "key1",
-                        ImageSrc = imageSrc,
-                        ImageAlt = imageAlt,
-                        ImageSize = new Size(width,height)
-                    }
-                 });
-             });
+        var component = RenderComponent<BitChoiceGroup>(parameters =>
+        {
+            parameters.Add(p => p.Options, new List<BitChoiceGroupOption>
+            {
+                new()
+                {
+                    Value = "key1",
+                    ImageSrc = imageSrc,
+                    ImageAlt = imageAlt,
+                    ImageSize = new Size(width,height)
+                }
+            });
+        });
 
         var bitChoiceGroupOption = component.Find(".bit-chgo");
-        Assert.IsTrue(bitChoiceGroupOption.ClassList.Contains($"bit-chgo-with-img-fluent"));
+        Assert.IsTrue(bitChoiceGroupOption.ClassList.Contains($"bit-chgo-with-img"));
 
         var image = component.Find(".bit-chgo-img img");
         Assert.IsTrue(image.HasAttribute("src"));
@@ -151,42 +146,44 @@ public class BitChoiceGroupTests : BunitTestContext
     [DataTestMethod, DataRow(BitIconName.Emoji2)]
     public void BitChoiceGroupShouldRespectIcon(BitIconName iconName)
     {
-        var component = RenderComponent<BitChoiceGroup>(
-            parameters =>
+        var component = RenderComponent<BitChoiceGroup>(parameters =>
+        {
+            parameters.Add(p => p.Options, new List<BitChoiceGroupOption>
             {
-                parameters.Add(p => p.Options, new List<BitChoiceGroupOption>
+                new BitChoiceGroupOption
                 {
-                    new BitChoiceGroupOption
-                    {
-                        Value = "key1",
-                        IconName = iconName
-                    }
-                });
+                    Value = "key1",
+                    IconName = iconName
+                }
             });
+        });
 
         var icon = component.Find(".bit-chgo-icon-wrapper i");
         Assert.IsTrue(icon.ClassList.Contains($"bit-icon--{iconName.GetName()}"));
     }
 
-    [DataTestMethod, DataRow(true, true), DataRow(true, false), DataRow(false, true), DataRow(false, false)]
-    public void BitChoiceGroupOtionOnChangeShouldWorkIfIsEnabled(bool groupIsEnabled, bool optionIsEnabled)
+    [DataTestMethod, 
+        DataRow(true, true),
+        DataRow(true, false),
+        DataRow(false, true),
+        DataRow(false, false)
+    ]
+    public void BitChoiceGroupOnChangeShouldWorkIfIsEnabled(bool groupIsEnabled, bool optionIsEnabled)
     {
         bool optionOnChangeValue = false;
 
-        var component = RenderComponent<BitChoiceGroup>(
-           parameters =>
-           {
-               parameters.Add(p => p.IsEnabled, groupIsEnabled);
-               parameters.Add(p => p.Options, new List<BitChoiceGroupOption>
-               {
-                    new()
-                    {
-                        Value = "key1",
-                        IsEnabled = optionIsEnabled,
-                        OnChange = () => optionOnChangeValue = true
-                    }
-               });
-           });
+        var component = RenderComponent<BitChoiceGroup>(parameters =>
+        {
+            parameters.Add(p => p.IsEnabled, groupIsEnabled);
+            parameters.Add(p => p.Options, new List<BitChoiceGroupOption>
+            {
+                new()
+                {
+                    Value = "key1",
+                    IsEnabled = optionIsEnabled
+                }
+            });
+        });
 
         var bitChoiceGroupOption = component.Find(".bit-chgo-input");
 
