@@ -40,7 +40,7 @@
     public int ClickedValue { get; set; }
     public int SelectedDateValue { get; set; }
 
-    public void HandleClick(MouseEventArgs e)
+    public void HandleClick()
     {
         ClickedValue++;
     }
