@@ -36,7 +36,7 @@
     public int ValidCount { get; set; }
     public int InvalidCount { get; set; }
 
-    public void HandleClick(MouseEventArgs e)
+    public void HandleClick()
     {
         ClickedValue++;
     }
