@@ -42,7 +42,7 @@
                    @onfocusin="@HandleFocusIn"
                    @onfocusout="@HandleFocusOut"
                    @onfocus="@HandleFocus"
-                   @onchange="@HandleChange" />
+                   @oninput="@HandleChange" />
 
             <i class="bit-icon bit-icon--@BitIconName.CalendarMirrored.GetName()" aria-hidden="true"></i>
         </div>
