@@ -3,8 +3,8 @@ namespace Bit.BlazorUI;
 
 internal static class BitOtpInputJsExtensions
 {
-    internal static async Task SetupOtpInputPaste(this IJSRuntime jsRuntime, DotNetObjectReference<BitOtpInput> obj, ElementReference otp)
+    internal static async Task SetupOtpInput(this IJSRuntime jsRuntime, DotNetObjectReference<BitOtpInput> obj, ElementReference otp)
     {
-        await jsRuntime.InvokeVoidAsync("BitOtpInput.setupOtpInputPaste", obj, otp);
+        await jsRuntime.InvokeVoidAsync("BitOtpInput.setupOtpInput", obj, otp);
     }
 }
