@@ -7,7 +7,7 @@
      class="@ClassBuilder.Value">
     <div class="bit-rbo-wrapper">
         <input @attributes="RadioButtonGroup?.InputHtmlAttributes"
-               id="@InputId"
+               id="@_inputId"
                class="bit-rbo-input"
                aria-label="@AriaLabel"
                type="radio"
@@ -15,44 +15,53 @@
                name="@Name"
                value="@Value"
                hidden
-               required="@(RadioButtonGroup?.IsRequired ?? false)"
+               required="@(RadioButtonGroup?.IsRequired)"
                @onclick="@HandleClick"
                @onchange="@HandleChange" />
 
-        <label for="@InputId" class="@GetLabelClassNameStr()">
-            @if (ImageSrc.HasValue())
+        <label for="@_inputId">
+            @if (ChildContent is not null)
             {
-                <div class="bit-rbo-img-inner-field">
-                    <div class="bit-rbo-img-wrapper">
-                        <div class="bit-rbo-img" style="@imageSizeStyle">
-                            <img src="@(isChecked ? SelectedImageSrc : ImageSrc)" alt="@ImageAlt">
-                        </div>
-                    </div>
-                </div>
-            }
-
-            @if (IconName.HasValue)
-            {
-                <div class="bit-rbo-icon-inner-field">
-                    <div class="bit-rbo-icon-wrapper">
-                        <i aria-hidden="true" class="bit-icon bit-icon--@IconName.GetName()"></i>
-                    </div>
-                </div>
-            }
-
-            @if (ImageSrc.HasValue() || IconName.HasValue)
-            {
-                <div class="bit-rbo-txt-wrapper">
-                    <span class="bit-rbo-txt" id="@TextId">
-                        @Text
-                    </span>
-                </div>
+                @ChildContent
             }
             else
             {
-                <span class="bit-rbo-txt" id="@TextId">
-                    @Text
-                </span>
+                <div class="@GetLabelClassNameStr()">
+                    @if (ImageSrc.HasValue())
+                    {
+                        <div class="bit-rbo-img-inner-field">
+                            <div class="bit-rbo-img-wrapper">
+                                <div class="bit-rbo-img" style="@_imageSizeStyle">
+                                    <img src="@(IsChecked ? SelectedImageSrc : ImageSrc)" alt="@ImageAlt">
+                                </div>
+                            </div>
+                        </div>
+                    }
+
+                    @if (IconName.HasValue)
+                    {
+                        <div class="bit-rbo-icon-inner-field">
+                            <div class="bit-rbo-icon-wrapper">
+                                <i aria-hidden="true" class="bit-icon bit-icon--@IconName.GetName()"></i>
+                            </div>
+                        </div>
+                    }
+
+                    @if (ImageSrc.HasValue() || IconName.HasValue)
+                    {
+                        <div class="bit-rbo-txt-wrapper">
+                            <span class="bit-rbo-txt" id="@_textId">
+                                @Text
+                            </span>
+                        </div>
+                    }
+                    else
+                    {
+                        <span class="bit-rbo-txt" id="@_textId">
+                            @Text
+                        </span>
+                    }
+                </div>
             }
         </label>
     </div>
