@@ -1,4 +1,6 @@
 ﻿@page "/components/drop-down"
+@inject HttpClient HttpClient
+@inject NavigationManager NavManager
 
 <ComponentDemo ComponentName="DropDown" ComponentDescription="A dropdown is a list in which the selected item is always visible while other items are visible on demand by clicking a dropdown button. Dropdowns are typically used for forms."
                ComponentParameters="componentParameters" EnumParameters="enumParameters">
@@ -9,20 +11,20 @@
                 <BitDropDown Label="Basic Uncontrolled"
                              Items="GetDropdownItems()"
                              Placeholder="Select an option"
-                             Style="width: 290px; margin: 20px 0 20px 0">
+                             Style="width: 100%; max-width: 290px; margin: 20px 0 20px 0">
                 </BitDropDown>
                 <BitDropDown Label="Disabled with defaultValue"
                              Items="GetDropdownItems()"
                              Placeholder="Select an option"
                              IsEnabled="false"
                              DefaultValue="v-bro"
-                             Style="width: 290px; margin-bottom: 20px;">
+                             Style="width: 100%; max-width: 290px; margin-bottom: 20px;">
                 </BitDropDown>
                 <BitDropDown Label="Multi-select uncontrolled"
                              Items="GetDropdownItems()"
                              Placeholder="Select options"
                              IsMultiSelect="true"
-                             Style="width: 290px; margin-bottom: 20px;">
+                             Style="width: 100%; max-width: 290px; margin-bottom: 20px;">
                 </BitDropDown>
             </div>
         </ExamplePreview>
@@ -34,7 +36,7 @@
                          Items="GetDropdownItems()"
                          Placeholder="Select an option"
                          @bind-Value="ControlledValue"
-                         Style="width: 290px; margin: 20px 0 20px 0">
+                         Style="width: 100%; max-width: 290px; margin: 20px 0 20px 0">
             </BitDropDown>
         </ExamplePreview>
     </ComponentExampleBox>
@@ -46,7 +48,7 @@
                          Placeholder="Select options"
                          @bind-Values="ControlledValues"
                          IsMultiSelect="true"
-                         Style="width:290px; margin:20px 0 20px 0">
+                         Style="width: 100%; max-width: 290px; margin:20px 0 20px 0">
             </BitDropDown>
         </ExamplePreview>
     </ComponentExampleBox>
@@ -57,7 +59,7 @@
                          Items="GetCustomDropdownItems()"
                          Placeholder="Select an option"
                          AriaLabel="Custom dropdown"
-                         Style="width:290px; margin:20px 0 20px 0">
+                         Style="width: 100%; max-width: 290px; margin:20px 0 20px 0">
                 <TextTemplate>
                     <div class="dropdown-customized-style">
                         <i class="bit-icon bit-icon--@((context.Items.Find(i => i.Value == context.Value).Data as DropDownItemData).IconName)"
@@ -89,7 +91,7 @@
                          Placeholder="Select an option"
                          Label="Custom Label"
                          AriaLabel="Custom dropdown label "
-                         Style="width:290px">
+                         Style="width: 100%; max-width: 290px">
                 <LabelFragment>
                     <label>Custom label</label>
                     <button type="button" title="Info" aria-label="Info" class="custom-drp-lbl-ic">
@@ -108,7 +110,7 @@
                                  Items="Categories"
                                  Placeholder="Select options"
                                  @bind-Value="@CurrentCategory"
-                                 Style="width:290px; margin:20px 0 20px 0">
+                                 Style="width: 100%; max-width: 290px; margin:20px 0 20px 0">
                     </BitDropDown>
 
                     <BitDropDown Label="Product"
@@ -116,7 +118,7 @@
                                  Placeholder="Select options"
                                  @bind-Value="@CurrentProduct"
                                  IsEnabled="string.IsNullOrEmpty(CurrentCategory) is false"
-                                 Style="width:290px; margin:20px 0 20px 0">
+                                 Style="width: 100%; max-width: 290px; margin:20px 0 20px 0">
                     </BitDropDown>
                 </div>
 
@@ -145,7 +147,7 @@
                                  IsMultiSelect="false"
                                  @bind-Value="formValidationDropDownModel.Category"
                                  Placeholder="Select an option"
-                                 Style="width: 290px; margin: 20px 0 20px 0" />
+                                 Style="width: 100%; max-width: 290px; margin: 20px 0 20px 0" />
 
                         <ValidationMessage For="@(() => formValidationDropDownModel.Category)" />
                     </div>
@@ -156,7 +158,7 @@
                                  IsMultiSelect="true"
                                  @bind-Values="formValidationDropDownModel.Products"
                                  Placeholder="Select an option"
-                                 Style="width: 290px; margin: 20px 0 20px 0" />
+                                 Style="width: 100%; max-width: 290px; margin: 20px 0 20px 0" />
 
                         <ValidationMessage For="@(() => formValidationDropDownModel.Products)" />
                     </div>
@@ -185,7 +187,7 @@
                              Items="GetDropdownItems()"
                              Placeholder="Select an option"
                              IsResponsiveModeEnabled=true
-                             Style="width: 290px; margin: 20px 0 20px 0">
+                             Style="width: 100%; max-width: 290px; margin: 20px 0 20px 0">
                 </BitDropDown>
             </div>
         </ExamplePreview>
@@ -200,7 +202,7 @@
                              IsResponsiveModeEnabled="true"
                              ShowSearchBox="true"
                              SearchBoxPlaceholder="Search item"
-                             Style="width: 290px; margin: 20px 0 20px 0">
+                             Style="width: 100%; max-width: 290px; margin: 20px 0 20px 0">
                 </BitDropDown>
 
                 <BitDropDown Label="Multi-select controlled with search box"
@@ -210,7 +212,7 @@
                              IsResponsiveModeEnabled="true"
                              ShowSearchBox="true"
                              SearchBoxPlaceholder="Search items"
-                             Style="width: 290px; margin: 20px 0 20px 0">
+                             Style="width: 100%; max-width: 290px; margin: 20px 0 20px 0">
                 </BitDropDown>
             </div>
         </ExamplePreview>
@@ -225,7 +227,7 @@
                          IsResponsiveModeEnabled="true"
                          ShowSearchBox="true"
                          SearchBoxPlaceholder="Search item"
-                         Style="width: 290px; margin: 20px 0 20px 0">
+                         Style="width: 100%; max-width: 290px; margin: 20px 0 20px 0">
             </BitDropDown>
 
             <BitDropDown Label="Multi-select controlled with virtualization"
@@ -236,7 +238,56 @@
                          IsResponsiveModeEnabled="true"
                          ShowSearchBox="true"
                          SearchBoxPlaceholder="Search items"
-                         Style="width: 290px; margin: 20px 0 20px 0">
+                         Style="width: 100%; max-width: 290px; margin: 20px 0 20px 0">
+            </BitDropDown>
+        </ExamplePreview>
+    </ComponentExampleBox>
+
+    <ComponentExampleBox Title="DropDown with virtualization (ItemsProvider)" HTMLSourceCode="@example10HTMLCode" CSharpSourceCode="@example10CSharpCode" ExampleId="example9">
+        <ExamplePreview>
+            <BitDropDown Label="Single-select Controlled with virtualization"
+                         ItemsProvider="LoadDropdownItems"
+                         Virtualize="true"
+                         Placeholder="Select an option"
+                         IsResponsiveModeEnabled="true"
+                         ShowSearchBox="true"
+                         SearchBoxPlaceholder="Search item"
+                         Style="width: 100%; max-width: 290px; margin: 20px 0 20px 0">
+            </BitDropDown>
+
+            <BitDropDown Label="Multi-select controlled with virtualization"
+                         ItemsProvider="LoadDropdownItems"
+                         Virtualize="true"
+                         Placeholder="Select options"
+                         IsMultiSelect="true"
+                         IsResponsiveModeEnabled="true"
+                         ShowSearchBox="true"
+                         SearchBoxPlaceholder="Search items"
+                         Style="width: 100%; max-width: 290px; margin: 20px 0 20px 0">
+            </BitDropDown>
+        </ExamplePreview>
+    </ComponentExampleBox>
+
+    <ComponentExampleBox Title="BitDropDown RTL" HTMLSourceCode="@example11HTMLCode" CSharpSourceCode="@example11CSharpCode" ExampleId="example11">
+        <ExamplePreview>
+            <BitDropDown Label="Single-select with Rtl direction"
+                         Items="GetArabicDropdownItems()"
+                         Placeholder="حدد اختيارا"
+                         IsResponsiveModeEnabled="true"
+                         IsRtl="true"
+                         ShowSearchBox="true"
+                         SearchBoxPlaceholder="عناصر البحث"
+                         Style="width: 100%; max-width: 290px; margin: 20px 0 20px 0">
+            </BitDropDown>
+            <BitDropDown Label="Multi-select with Rtl direction"
+                         Items="GetArabicDropdownItems()"
+                         Placeholder="اشر على الخيارات"
+                         IsMultiSelect="true"
+                         IsResponsiveModeEnabled="true"
+                         IsRtl="true"
+                         ShowSearchBox="true"
+                         SearchBoxPlaceholder="عناصر البحث"
+                         Style="width: 100%; max-width: 290px; margin-bottom: 20px;">
             </BitDropDown>
         </ExamplePreview>
     </ComponentExampleBox>
