@@ -8,6 +8,7 @@ namespace Bit.BlazorUI.Playground.Web.Pages.Components.RadioButtonGroup;
 public partial class BitRadioButtonGroupDemo
 {
     private string MyValue = "B";
+    private string CustomOptionValue = "Day";
     private string SuccessMessage = string.Empty;
     private FormValidationModel FormValidationModel = new();
 
@@ -140,7 +141,8 @@ public partial class BitRadioButtonGroupDemo
 
     #region Example Code 1
 
-    private readonly string example1HTMLCode = @"<BitLabel>Selected Key is : @MyValue</BitLabel>
+    private readonly string example1HTMLCode = @"
+<BitLabel>Selected Key is : @MyValue</BitLabel>
 <BitTextField @bind-Value=""MyValue"" Placeholder=""Select one of A, B or C""></BitTextField>
 <BitRadioButtonGroup Name=""Group1"" Label=""Pick one"" IsRequired=""true"" @bind-Value=""MyValue"">
     <BitRadioButtonOption Text=""Option A"" Value=""A""></BitRadioButtonOption>
@@ -155,7 +157,8 @@ private string MyValue = ""B"";";
 
     #region Example Code 2
 
-    private readonly string example2HTMLCode = @"<BitRadioButtonGroup Name=""Group2"" IsEnabled=""false"" Label=""Pick one"" DefaultValue=""C"">
+    private readonly string example2HTMLCode = @"
+<BitRadioButtonGroup Name=""Group2"" IsEnabled=""false"" Label=""Pick one"" DefaultValue=""C"">
     <BitRadioButtonOption Text=""Option A"" Value=""A""></BitRadioButtonOption>
     <BitRadioButtonOption Text=""Option2 B"" Value=""B""></BitRadioButtonOption>
     <BitRadioButtonOption Text=""Disabled option C"" Value=""C"" IsEnabled=""false""></BitRadioButtonOption>
@@ -165,7 +168,8 @@ private string MyValue = ""B"";";
 
     #region Example Code 3
 
-    private readonly string example3HTMLCode = @"<BitRadioButtonGroup Name=""Group1"" Label=""Pick one image"" DefaultValue=""pie"">
+    private readonly string example3HTMLCode = @"
+<BitRadioButtonGroup Name=""Group1"" Label=""Pick one image"" DefaultValue=""pie"">
     <BitRadioButtonOption Text=""Clustered bar chart"" Value=""bar"" ImageSrc=""https://static2.sharepointonline.com/files/fabric/office-ui-fabric-react-assets/choicegroup-bar-unselected.png"" ImageAlt=""alt for image Option 1"" SelectedImageSrc=""https://static2.sharepointonline.com/files/fabric/office-ui-fabric-react-assets/choicegroup-bar-selected.png"" ImageSize=""new System.Drawing.Size( width: 32, height: 32)""></BitRadioButtonOption>
     <BitRadioButtonOption Text=""Pie chart"" Value=""pie"" ImageSrc=""https://static2.sharepointonline.com/files/fabric/office-ui-fabric-react-assets/choicegroup-bar-unselected.png"" ImageAlt=""alt for image Option 2"" SelectedImageSrc=""https://static2.sharepointonline.com/files/fabric/office-ui-fabric-react-assets/choicegroup-bar-selected.png"" ImageSize=""new System.Drawing.Size( width: 32, height: 32)""></BitRadioButtonOption>
     <BitRadioButtonOption Text=""Disabeled"" IsEnabled=""false"" Value=""disabeled_option"" ImageSrc=""https://static2.sharepointonline.com/files/fabric/office-ui-fabric-react-assets/choicegroup-bar-unselected.png"" ImageAlt=""alt for image Option 2"" SelectedImageSrc=""https://static2.sharepointonline.com/files/fabric/office-ui-fabric-react-assets/choicegroup-bar-selected.png"" ImageSize=""new System.Drawing.Size( width: 32, height: 32)""></BitRadioButtonOption>
@@ -175,7 +179,8 @@ private string MyValue = ""B"";";
 
     #region Example Code 4
 
-    private readonly string example4HTMLCode = @"<BitRadioButtonGroup Name=""Group1"" Label=""Pick one icon"">
+    private readonly string example4HTMLCode = @"
+<BitRadioButtonGroup Name=""Group1"" Label=""Pick one icon"">
     <BitRadioButtonOption Text=""Day"" Value=""day"" IconName=""BitIconName.CalendarDay""></BitRadioButtonOption>
     <BitRadioButtonOption Text=""Week"" Value=""week"" IconName=""BitIconName.CalendarWeek""></BitRadioButtonOption>
     <BitRadioButtonOption Text=""Month"" Value=""month"" IconName=""BitIconName.Calendar"" IsEnabled=""false""></BitRadioButtonOption>
@@ -185,7 +190,76 @@ private string MyValue = ""B"";";
 
     #region Example Code 5
 
-    private readonly string example5HTMLCode = @"@if (string.IsNullOrEmpty(SuccessMessage))
+    private readonly string example5HTMLCode = @"
+<style>
+    .custom-option {
+        display: flex;
+        align-items: center;
+        gap: rem(10px);
+        cursor: pointer;
+
+        .radio-pointer {
+            width: rem(20px);
+            height: rem(20px);
+            border: 1px solid;
+            border-radius: rem(10px);
+        }
+
+        &:hover {
+            .radio-pointer {
+                border-top: rem(5px) solid #C66;
+                border-bottom: rem(5px) solid #6C6;
+                border-left: rem(5px) solid #66C;
+                border-right: rem(5px) solid #CC6;
+            }
+        }
+
+        &.selected-option {
+            color: #C66;
+
+            .radio-pointer {
+                border-top: rem(10px) solid #C66;
+                border-bottom: rem(10px) solid #6C6;
+                border-left: rem(10px) solid #66C;
+                border-right: rem(10px) solid #CC6;
+            }
+        }
+    }
+</style>
+
+<BitRadioButtonGroup Name=""Group1"" Label=""Pick one option"" @bind-Value=""CustomOptionValue"">
+    <BitRadioButtonOption Value=""Day"">
+        <div class=""custom-option @(CustomOptionValue is ""Day"" ? ""selected-option"" : """")"">
+            <div class=""radio-pointer""></div>
+            <BitIcon IconName=""BitIconName.CalendarDay"" />
+            <span>Day</span>
+        </div>
+    </BitRadioButtonOption>
+    <BitRadioButtonOption Value=""Week"">
+        <div class=""custom-option @(CustomOptionValue is ""Week"" ? ""selected-option"" : """")"">
+            <div class=""radio-pointer""></div>
+            <BitIcon IconName=""BitIconName.CalendarWeek"" />
+            <span>Week</span>
+        </div>
+    </BitRadioButtonOption>
+    <BitRadioButtonOption Value=""Month"">
+        <div class=""custom-option @(CustomOptionValue is ""Month"" ? ""selected-option"" : """")"">
+            <div class=""radio-pointer""></div>
+            <BitIcon IconName=""BitIconName.Calendar"" />
+            <span>Month</span>
+        </div>
+    </BitRadioButtonOption>
+</BitRadioButtonGroup>";
+
+    private readonly string example5CSharpCode = @"
+private string CustomOptionValue = ""Day"";";
+
+    #endregion
+
+    #region Example Code 6
+
+    private readonly string example6HTMLCode = @"
+@if (string.IsNullOrEmpty(SuccessMessage))
 {
     <EditForm Model=""FormValidationModel"" OnValidSubmit=""HandleValidSubmit"" OnInvalidSubmit=""HandleInvalidSubmit"">
         <DataAnnotationsValidator />
@@ -216,7 +290,8 @@ else
     </BitMessageBar>
 }";
 
-    private readonly string example5CSharpCode = @"public class FormValidationModel
+    private readonly string example6CSharpCode = @"
+public class FormValidationModel
 {
     [Required]
     public string Option { get; set; }
