@@ -29,4 +29,4 @@
     </div>
 </div>
 
-<NavigationButtons Prev="Exception handling" PrevUrl="/todo-template/exception-handling" Next="Contribute" NextUrl="/todo-template/contribute"/>
+<NavigationButtons Prev="Exception handling" PrevUrl="/todo-template/exception-handling" Next="Multilingualism" NextUrl="/todo-template/multilingualism" />
