@@ -3,7 +3,7 @@
 <div class="page-container">
     <h1 class="page-main-title">Contribute</h1>
     <div class="page-desc">
-        There are many ways to contribute in the Bit TodoTemplate project:
+        There are many ways to contribute in the Bit platform projects:
     </div>
     <div class="section-card">
         <div class="section-card-txt">
@@ -24,4 +24,4 @@
     </div>
 </div>
 
-<NavigationButtons Prev="Cache mechanism" PrevUrl="/todo-template/cache-mechanism"/>
+<NavigationButtons Prev="DevOps" PrevUrl="/todo-template/dev-ops" />
