@@ -35,8 +35,8 @@
     <section class="section-card">
         <h2 class="section-card-title">Migration</h2>
         <div class="section-card-txt">To create and migrate the database to the latest version. You can use Entity Framework's built-in tools for migrations. Open <span class="bold-txt">Package Manager Console</span> in Visual Studio set <span class="bold-txt">TodoTemplate.Server.Api</span> as the Default Project and run the Update-Database command as shown below:</div>
-        <div class="code-box"><pre><code>Update-Database -Context AppDbContext</code></pre></div>
+        <div class="code-box"><pre><code>Update-Database</code></pre></div>
     </section>
 </div>
 
-<NavigationButtons Prev="Getting started" PrevUrl="/todo-template/getting-started" Next="Run" NextUrl="/todo-template/run"/>
+<NavigationButtons Prev="Getting started" PrevUrl="/todo-template/getting-started" Next="Run" NextUrl="/todo-template/run" />
