@@ -45,4 +45,4 @@
     </section>
 </div>
 
-<NavigationButtons Prev="Hosting models" PrevUrl="/todo-template/hosting-models" Next="Settings" NextUrl="/todo-template/settings"/>
+<NavigationButtons Prev="Hosting models" PrevUrl="/todo-template/hosting-models" Next="Settings" NextUrl="/todo-template/settings" />
