@@ -24,7 +24,7 @@
         <div class="section-card-txt">
             <ul>
                 <li>
-                    Microsoft Visual Studio 2022 - Preview Version 17.3.0 Preview or higher with the following workloads and extention
+                    Microsoft Visual Studio 2022 - Preview Version 17.4.0 or higher with the following workloads and extention
                     <ul dir="auto">
                         <li>ASP.NET and web development</li>
                         <li>Net Multi-Platform App UI development</li>
