@@ -37,4 +37,4 @@
     </section>
 </div>
 
-<NavigationButtons Prev="Project structure" PrevUrl="/todo-template/project-structure" Next="Cache mechanism" NextUrl="/todo-template/cache-mechanism"/>
+<NavigationButtons Prev="Project structure" PrevUrl="/todo-template/project-structure" Next="Cache mechanism" NextUrl="/todo-template/cache-mechanism" />
