@@ -10,7 +10,7 @@
         <div class="section-card-txt">
             1. Install the TodoTemplate Template from the NuGet package manager.
         </div>
-        <div class="code-box">dotnet new --install Bit.TodoTemplate</div>
+        <div class="code-box">dotnet new install Bit.TodoTemplate</div>
         <div class="section-card-txt">
             2. Create a new project using the Template installed
         </div>
@@ -19,12 +19,8 @@
             Or create a new project with the Visual Studio.
         </div>
         <div class="image-container">
-            <img class="image" alt="Create new project" width="1010" height="673" src="/images/demo/create-project.webp" />
+            <img class="image" alt="Create new project" width="1010" height="673" src="/images/demo/create-project-todo.webp" />
         </div>
-        <div class="section-card-txt">
-            3. Make sure the first line in the <span class="high-light-text">TodoTemplate.Client.Web.csproj</span> file contains the following value.
-        </div>
-        <div class="code-box">@("<Project Sdk=\"Microsoft.NET.Sdk.Web\">")</div>
         <div class="section-card-txt"><span class="bold-txt">Note</span>: By default, the current folder name will be considered as the solution name. If you want to choose another name, use the switch -n as below.</div>
         <div class="code-box">dotnet new Bit.TodoTemplate -n [YourSolutionName]</div>
     </section>
