@@ -1,7 +1,7 @@
 ﻿@page "/todo-template/hosting-models"
 
 <div class="page-container">
-    <h1 class="page-main-title">Hosting models (Blazoe Mode)</h1>
+    <h1 class="page-main-title">Hosting models (Blazor Mode)</h1>
     <div class="page-desc">
         Bit TodoTemplate uses Blazor for building UI. Blazor is a web framework for building web UI components 
         (<a href="https://docs.microsoft.com/en-us/aspnet/core/blazor" title="ASP.NET Core Blazor" rel="nofollow">ASP.NET Core Blazor</a>)
@@ -84,9 +84,13 @@
             <img class="image" alt="Debug target" width="1014" height="350" src="/images/demo/debug-target.webp" />
         </div>
         <div>
-            3. Set solution on Multi-startup project, by right click on solution name and selecting properties from right-click menu (change the action of the Api, App projects to start)
+            3. Change <span class="bold-txt">ApiServerAddress</span> in Client/Shared/wwwroot/appsettings.json to make sure your Android / iOS device can reach the server.
+        </div>
+        <br />
+        <div>
+            4. Set solution on Multi-startup project, by right click on solution name and selecting properties from right-click menu (change the action of the Api, App projects to start)
         </div>
     </section>
 </div>
 
-<NavigationButtons Prev="Run" PrevUrl="/todo-template/run" Next="Deployment type" NextUrl="/todo-template/deployment-type"/>
+<NavigationButtons Prev="Run" PrevUrl="/todo-template/run" Next="Deployment type" NextUrl="/todo-template/deployment-type" />
