@@ -15,12 +15,8 @@
         </div>
 
         <div class="section-card-txt">
-            <span class="bold-txt">TodoTemplate.Server.Api</span>:
+            <span class="bold-txt">TodoTemplate.Api</span>:
             This project contains server-side codes including Controllers, Data access layer (EntityFramework ORM), Identity, etc.
-        </div>
-        <div class="section-card-txt">
-            <span class="bold-txt">TodoTemplate.Client.App</span>:
-            This project contains code related to building an app, although much of the code is common to the web project and refers to it.
         </div>
         <div class="section-card-txt">
             <span class="bold-txt">TodoTemplate.Iac</span>:
@@ -28,13 +24,21 @@
             You don't have to use either Pulumi or even Azure, Host it wherever you want.
         </div>
         <div class="section-card-txt">
-            <span class="bold-txt">TodoTemplate.Shared</span>: This project is shared on both the server (Api project) and client(Web, App projects) sides, and can share code. The bit puts the code it wants to reuse; Like resources, Dtos, shared contracts, services, etc.
+            <span class="bold-txt">TodoTemplate.Shared</span>: This project is shared on both the server (Api project) and client(Web, App projects) sides, and can share code. The bit puts the code it wants to reuse; Like resources, Dto classes, shared contracts, services, etc.
         </div>
         <div class="section-card-txt">
-            <span class="bold-txt">TodoTemplate.Client.Web</span>: This project contains the necessary code to build the project UI, whether it is a web output or an application. And includes components and pages made with the help of Razor and Style sheet
+            <span class="bold-txt">TodoTemplate.App</span>:
+            This project contains code related to building an app, although much of the code is common to the web project which are located in TodoTemplate.Client.Shared.
+        </div>
+        <div class="section-card-txt">
+            <span class="bold-txt">TodoTemplate.Client.Shared</span>:
+            This project includes components and pages made with the help of Razor and Style sheet
+        </div>
+        <div class="section-card-txt">
+            <span class="bold-txt">TodoTemplate.Web</span>: This project contains the necessary code to run web version of the project (Blazor WebAssembly & Blazor Server)
         </div>
 
     </section>
 </div>
 
-<NavigationButtons Prev="Settings" PrevUrl="/todo-template/settings" Next="Exception handling" NextUrl="/todo-template/exception-handling"/>
+<NavigationButtons Prev="Settings" PrevUrl="/todo-template/settings" Next="Exception handling" NextUrl="/todo-template/exception-handling" />
