@@ -112,4 +112,4 @@
     </section>
 </div>
 
-<NavigationButtons Prev="Deployment type" PrevUrl="/todo-template/deployment-type" Next="Project structure" NextUrl="/todo-template/project-structure"/>
+<NavigationButtons Prev="Deployment type" PrevUrl="/todo-template/deployment-type" Next="Project structure" NextUrl="/todo-template/project-structure" />
