@@ -60,8 +60,32 @@
         a {
             text-decoration: none;
         }
+
+        ::-webkit-scrollbar {
+            width: 6.5px;
+            background-color: #fff;
+        }
+
+        ::-webkit-scrollbar-track {
+            background-color: #fff;
+        }
+
+        ::-webkit-scrollbar-thumb {
+            background-color: #c8c6c4;
+            border-radius: 8px;
+            border: 4px solid #c8c6c4;
+        }
+
+        ::-webkit-scrollbar-button {
+            display: none;
+        }
+
+        ::-webkit-scrollbar-thumb:hover {
+            background-color: #0064EF;
+            border-radius: 8px;
+            border: 4px solid #0064EF;
+        }
     </style>
-    @*<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.3.1/styles/default.min.css">*@
     <component type="typeof(HeadOutlet)" render-mode="@renderMode" />
 </head>
 <body>
@@ -95,8 +119,13 @@
                     'https://www.googletagmanager.com/gtm.js?id=' + i + dl; f.parentNode.insertBefore(j, f);
         })(window, document, 'script', 'dataLayer', '@googleTagContainerID');</script>
     <!-- End Google Tag Manager -->
+    
     <!--BEGIN RAYCHAT CODE-->
     <script type="text/javascript">!function () { function t() { var t = document.createElement("script"); t.type = "text/javascript", t.async = !0, localStorage.getItem("rayToken") ? t.src = "https://app.raychat.io/scripts/js/" + o + "?rid=" + localStorage.getItem("rayToken") + "&href=" + window.location.href : t.src = "https://app.raychat.io/scripts/js/" + o + "?href=" + window.location.href; var e = document.getElementsByTagName("script")[0]; e.parentNode.insertBefore(t, e) } var e = document, a = window, o = "@rayChatToken"; "complete" == e.readyState ? t() : a.attachEvent ? a.attachEvent("onload", t) : a.addEventListener("load", t, !1) }();</script>
     <!--END RAYCHAT CODE-->
+
+    <!-- TrustBox script -->
+    <script type="text/javascript" src="//widget.trustpilot.com/bootstrap/v5/tp.widget.bootstrap.min.js" async></script>
+    <!-- End TrustBox script -->
 </body>
 </html>
