@@ -4,23 +4,30 @@
             <div class="bit-logo"></div>
             <div class="footer-desc">A platform that enables freelancers to Code in a clean project template and Build & Deploy with 1-click.</div>
         </div>
-        <div class="footer-col">
-            <div class="col-title">Products</div>
-            <a class="footer-link" target="_blank" href="@Urls.TodoTemplate">TodoTemplate</a>
-            <a class="footer-link" target="_blank" href="@Urls.AdminPanelTemplate">AdminPanel</a>
-            <a class="footer-link" target="_blank" href="@Urls.Components">Bit BlazorUI</a>
-            
-        </div>
-        <div class="footer-col">
-            <div class="col-title">Quick links</div>
-            <a class="footer-link" target="_blank" href="@Urls.GithubRepo">Github Repo</a>
-            <a class="footer-link" target="_blank" href="@Urls.GithubOrg">Github Org</a>
-            <a class="footer-link" target="_blank" href="@Urls.GithubIssues">Issues</a>
-        </div>
-        <div class="footer-col">
-            <div class="col-title">Media</div>
-            <a href="@Urls.Blog" target="_blank" class="footer-link">Blog</a>
-            <a href="@Urls.Video" target="_blank" class="footer-link">Video</a>
+        <div class="grid-container">
+            <div class="footer-col">
+                <div class="col-title">Products & Services</div>
+                <a class="footer-link" href="@Urls.TodoTemplate">TodoTemplate</a>
+                <a class="footer-link" href="@Urls.AdminPanel">AdminPanel</a>
+                <a class="footer-link" href="@Urls.Components">BlazorUI</a>
+                @*<a class="footer-link" href="@Urls.CloudHostingSolutins">Cloud hostring solutions</a>*@
+                @*<a class="footer-link" href="@Urls.Academy">Bit academy</a>*@
+
+            </div>
+            <div class="footer-col">
+                <div class="col-title">Community</div>
+                <a class="footer-link" target="_blank" href="@Urls.AboutUs">About</a>
+                <a class="footer-link" target="_blank" href="@Urls.GithubRepo">Github repo</a>
+                <a class="footer-link" target="_blank" href="@Urls.GithubOrg">Github org</a>
+                <a class="footer-link" target="_blank" href="@Urls.GithubIssues">Issues</a>
+                <a class="footer-link" target="_blank" href="@Urls.ContactUs">Contact us</a>
+            </div>
+
+            <div class="footer-col">
+                <div class="col-title">Media</div>
+                <a class="footer-link" href="#">Blogs</a>
+                <a class="footer-link" href="#">Videos</a>
+            </div>
         </div>
     </section>
 
